modify useAbility to be a regular function and move to utils, remove redundant enums, move fireArrowsinArea step to section file
